requirements.txt
3to2==1.1.1
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Unidecode==0.4.18
Werkzeug==0.10.4
argparse==1.2.1
cffi==1.2.1
cryptography==1.0
decorator==4.0.2
elasticsearch==1.6.0
enum34==1.0.4
fuzzywuzzy==0.6.1
idna==2.0
ipaddress==1.0.14
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
ndg-httpsclient==0.4.0
nltk==3.0.4
oauthlib==1.0.1
path.py==7.6.1
pexpect==3.3
pickleshare==0.5
pika==0.9.8
pyOpenSSL==0.15.1
pyasn1==0.1.8
pycparser==2.14
pyenchant==1.6.6
python-Levenshtein==0.12.0
python-nvd3==0.13.10
python-slugify==0.1.0
requests==2.7.0
requests-oauthlib==0.5.0
simplegeneric==0.8.1
six==1.9.0
textblob==0.9.1
traitlets==4.0.0
twython==3.3.0
urllib3==1.11
wsgiref==0.1.2
